[WIP] Move movie list dto
