+++ data/src/main/java/com/example/data/mapper/MoviesDtoMapper.kt
@@ -1,6 +1,6 @@
 package com.example.data.mapper
 
-import com.example.data.remote.dto.MoviesDto
+import com.example.data.remote.dto.movielist.MoviesDto
 import com.example.domain.model.MovieModel
 import com.example.domain.wrapper.Mapper
