+++ data/src/main/java/com/example/data/remote/dto/movielist/AgeRating.kt
@@ -1,4 +1,4 @@
-package com.example.data.remote.dto
+package com.example.data.remote.dto.movielist
 
 
 import kotlinx.serialization.SerialName
