+++ data/src/main/java/com/example/data/remote/dto/movielist/Images.kt
@@ -1,7 +1,6 @@
-package com.example.data.remote.dto
+package com.example.data.remote.dto.movielist
 
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
