+++ data/src/main/java/com/example/data/remote/dto/movielist/Poster.kt
@@ -1,11 +1,11 @@
-package com.example.data.remote.dto
+package com.example.data.remote.dto.movielist
 
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Still(
+data class Poster(
     @SerialName("1")
-    val x1: X1X
+    val x1: X1
 )
